requirements.txt
boto3
Pillow
botocore
